Make redundancy bugs all face the program now. Remove the stopwatch item from the bug redundancy class.

+++ BugSquashLib/Bug.h
@@ -26,7 +26,7 @@ private:
 	double mStartTime = 0;
 
 	/// The last stopwatch time
-	long mTime = 0;
+	double mTime = 0;
 
 	/// Stopwatch used to measure elapsed time
 	wxStopWatch mStopWatch;
